Use RequiresInstalledToolChain and merge spec class in Xcode test
