+++ subprojects/ide-native/src/integTest/groovy/org/gradle/ide/xcode/XcodeMultipleProjectIntegrationTest.groovy
@@ -17,10 +17,9 @@
 package org.gradle.ide.xcode
 
 import org.gradle.ide.xcode.fixtures.AbstractXcodeIntegrationSpec
-import org.gradle.ide.xcode.fixtures.SwiftToolChainTestingSpec
 import org.gradle.nativeplatform.fixtures.app.SwiftAppWithLibrary
 
-class XcodeMultipleProjectIntegrationTest extends AbstractXcodeIntegrationSpec implements SwiftToolChainTestingSpec {
+class XcodeMultipleProjectIntegrationTest extends AbstractXcodeIntegrationSpec {
     def setup() {
         settingsFile << """
             include 'app', 'greeter'
