+++ subprojects/ide-native/src/integTest/groovy/org/gradle/ide/xcode/XcodeMultipleSwiftProjectIntegrationTest.groovy
@@ -390,7 +390,7 @@ class XcodeMultipleSwiftProjectIntegrationTest extends AbstractXcodeIntegrationS
     }
 
     def "can create xcode project for Swift application inside composite build"() {
-        useSwiftCompiler()
+        requireSwiftToolChain()
 
         given:
         settingsFile.text = """
